% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redatam.R
\name{read_redatam}
\alias{read_redatam}
\title{Read REDATAM file}
\usage{
read_redatam(file_path)
}
\arguments{
\item{file_path}{The path to the REDATAM file.}
}
\value{
A character vector with the contents of the file.
}
\description{
This function reads a REDATAM file using the C++ code.
}
