% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-package.R
\docType{package}
\name{redatam-package}
\alias{redatam}
\alias{redatam-package}
\title{COMPLETE}
\description{
COMPLETE
}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
