% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam_structure.R
\name{redatam_structure}
\alias{redatam_structure}
\title{Extract the structure of a REDATAM .dicx file}
\usage{
redatam_structure(dicx_file)
}
\arguments{
\item{dicx_file}{character path to the .dicx file}
}
\description{
Extract the structure of a REDATAM .dicx file
}
