% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-package.R
\name{read_redatam}
\alias{read_redatam}
\title{Read a REDATAM dictionary file with the corresponding entity tables}
\usage{
read_redatam(dic_path)
}
\arguments{
\item{dic_path}{Path to the REDATAM dictionary file}
}
\description{
Read a REDATAM dictionary file with the corresponding entity tables
}
